requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.4.26
cffi==1.17.1
cloudinary==1.44.0
cryptography==44.0.2
dj-database-url==2.3.0
Django==4.2.20
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
importlib-metadata==8.5.0
Markdown==3.7
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
urllib3==2.2.3
zipp==3.20.2
